Fixxing delete location key violation

+++ backend/api/locations.js
@@ -82,6 +82,15 @@ export const locationsIdEdit = {
 			return res.status(404).json({ error: "Location not found" });
 		}
 
+		//check if used somewhere 
+		let bookingsCount = await db.countBookingsInLocation(req.params.id);
+
+		if (bookingsCount > 0) {
+            return res.status(409).json({
+                error: "Cannot change location with existing references",
+            });
+        }
+
 		let editedLocation = await db.editLocation(req.params.id, req.body.name);
 		res.json(editedLocation);
 	},
@@ -96,6 +105,15 @@ export const locationIdDelete = {
 		if (!location) {
 			return res.status(404).json({ error: "Location not found" });
 		}
+
+		//check if used somewhere 
+		let bookingsCount = await db.countBookingsInLocation(req.params.id);
+
+		if (bookingsCount > 0) {
+            return res.status(409).json({
+                error: "Cannot change location with existing references",
+            });
+        }
 
 		await db.deleteLocation(req.params.id);
 		res.sendStatus(204);
